requirements.txt
streamlit>=1.20.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
tqdm
spacy>=3.5.0
